PO-8916 Address review omments
